Grammar error
Found a grammar error line 620. Changed "That one file doesn't appear to have an extension. Strange. Take a look at more of the that "license" file that doesn't show an extension." to "That one file doesn't appear to have an extension. Strange. Take a look at **more** of the \"license\" file that doesn't show an extension."